requirements.txt
setuptools==49.2.1

cltl.brain==0.0.dev7
cltl.combot==0.0.dev1
cltl.asr[impl]==0.0.dev4
cltl.vad[impl]==0.0.dev1
cltl.backend[impl,host]==0.0.dev3
cltl.reply-generation==0.0.dev8
cltl.triple-extraction==0.0.dev8
emissor==0.0.dev5
jsonpickle==2.0.0
opencv-python==4.5.4.58
python_on_whales==0.31.0
pillow~=9.0.1
spacy~=3.2.3
tqdm
ipykernel~=6.9.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
requests~=2.27.1
seaborn==0.11.2
torch
torchaudio==0.9.1
torchvision==0.10.1
transformers==4.10.0
gTTS==2.2.3

jupyter~=1.0.0
PyYAML~=6.0
numpy~=1.22.2
pandas~=1.4.1
defusedxml~=0.7.1
cffi~=1.15.0
ipython~=8.1.0
pip~=21.3
attrs~=21.4.0
tornado~=6.1
Jinja2~=3.0.3
pyzmq~=22.3.0
pexpect~=4.8.0
blis~=0.7.6
nltk~=3.4.5
click~=8.0.4
six~=1.16.0
urllib3~=1.26.8
parso~=0.8.3
jedi~=0.18.1
scipy~=1.8.0
regex~=2022.3.2
pyparsing~=3.0.7
pytz~=2021.3
QtPy~=2.0.1
packaging~=21.3
ipywidgets~=7.6.5
notebook~=6.4.8
cymem~=2.0.6
MarkupSafe~=2.1.0
pathy~=0.6.1
typer~=0.4.0
thinc~=8.0.13
srsly~=2.4.2
wasabi~=0.9.0
pydantic~=1.8.2
catalogue~=2.0.6
langcodes~=3.3.0
preshed~=3.0.6
stanfordnlp~=0.2.0
murmurhash~=1.0.6
webencodings~=0.5.1
bleach~=4.1.0
joblib~=1.1.0
wget~=3.2
python-dateutil~=2.8.2
certifi~=2021.10.8
debugpy~=1.5.1
traitlets~=5.1.1
decorator~=5.1.1
Pygments~=2.11.2
testpath~=0.6.0
backcall~=0.2.0
pickleshare~=0.7.5
wcwidth~=0.2.5
ptyprocess~=0.7.0
cycler~=0.11.0
nbformat~=5.1.3
nbclient~=0.5.11
nbconvert~=6.4.2
jsonschema~=4.4.0
Send2Trash~=1.8.0
terminado~=0.13.1
idna~=3.3
asttokens~=2.0.5
zipp~=3.7.0
executing~=0.8.3
pandocfilters~=1.5.0
mistune~=0.8.4
entrypoints~=0.4
pycparser~=2.21
qtconsole~=5.2.2
pyrsistent~=0.18.1
kiwisolver~=1.3.2
sacremoses~=0.0.47
tokenizers~=0.10.3
filelock~=3.6.0
mdformat==0.7.14
flake8==4.0.1
pyflakes==2.4.0
isort==5.10.1
black==22.3.0